refactor(cvss): split versions in modules
