+++ code/cvss/cvss.cabal
@@ -15,6 +15,12 @@ tested-with:
 
 library
   exposed-modules:  Security.CVSS
+  other-modules:  Security.CVSS.Types
+                 , Security.CVSS.Internal
+                 , Security.CVSS.V20
+                 , Security.CVSS.V30
+                 , Security.CVSS.V31
+                 , Security.CVSS.V40
   build-depends:
     , base         >=4.14 && <5
     , containers    >=0.6   && <0.8
